feat: finish profile page fuctional
