+++ src/components/profile-orders/profile-orders.jsx
@@ -1,0 +1,3 @@
+const ProfileOrders = () => {};
+
+export default ProfileOrders;
