+++ src/pages/profile/profile.module.css
@@ -27,19 +27,3 @@
   composes: pt-20 text text_type_main-default text_color_inactive from global;
   opacity: 0.4;
 }
-
-.form {
-  display: flex;
-  flex-direction: column;
-  row-gap: 24px;
-}
-
-.buttons {
-  align-self: end;
-  display: flex;
-  column-gap: 20px;
-}
-
-.cancelButton {
-  padding: 16px 8px;
-}
